Further refinements and integration
