+++ static/.eslint-ce-static.yml
@@ -44,6 +44,9 @@ rules:
   no-useless-escape: error
   no-useless-rename: error
   no-useless-return: error
+  no-empty:
+    - error
+    - allowEmptyCatch: true
   quote-props:
     - error
     - as-needed
